evaluation/requirements.txt
easyocr
tqdm
python-levenshtein
torch
